+++ animory/src/test/java/com/daggle/animory/acceptance/PetTest.java
@@ -1,0 +1,2 @@
+package com.daggle.animory.acceptance;public class PetTest {
+}
